+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.6.3
 
+- FIX: Correctly handle [trailingSlash config option](https://docusaurus.io/docs/docusaurus.config.js/#trailing-slash) (#52)
 - FEAT: Fully localized autocomplete (currently in German, English, Italian, Brazilian Portuguese, and Chinese).
 - FEAT: Update autocomplete to 1.2.0
 - FEAT: Support Hindi for search indexing (`hi`)